+++ R/retrieve_qde.R
@@ -2,44 +2,44 @@
 #'
 #' This function download the FAERS data for a specific year and quarter
 #'
-#' @param year The year of the data to download
-#' @param quarter The quarter of the data to download
-#' @param type The format of the data to download, ascii or xml
+#' @param path (char) The path of the folder where the database
+#' will be saved (default: working directory)
+#' @param year (int) The year of the data to download (no default)
+#' @param quarter (char) The quarter of the data to download
+#' ("q1", "q2", "q3" or "q4", default: "q1)
+#' @param type (char) The format of the data to download,
+#' ("ascii" or "xml", default: "ascii")
 #'
-#' @return The function download the selected data in the working directory
+#' @return The function download the selected data
 #' @export
 #'
 #' @examples
-#' retrieve_qde(year = "2018", quarter = "q1", type = "ascii")
+#' retrieve_qde(year = 2018, quarter = "q4")
 #'
-retrieve_qde <- function(year = format(Sys.Date(), "%Y"),
+retrieve_qde <- function(path = getwd(),
+                         year,
                          quarter = c("q1", "q2", "q3", "q4"),
                          type = c("ascii", "xml")) {
-  # Path of the online data
-  sourcedir <- paste0("https://fis.fda.gov/content/Exports/faers_",
-                      type, "_", year, quarter, ".zip")
-
-  # Create the directory for the downloaded zip folder
-  destdir <- paste0("faers_raw_data/",
-                    year, "/", quarter)
-  dir.create(destdir, recursive = T)
-
-  # Create the path of the zipped folder
-  filepath <- paste0(getwd(), "/", destdir, "/", year, quarter, ".zip")
-
-  # Download the the zipped folder
-  downloader::download(sourcedir, filepath, mode = "wb")
-
-  # Unzip the folder
-  utils::unzip(zipfile = filepath,
-               exdir = paste0(getwd(), "/", destdir))
-
-  # Keep only the data files, delete pdf and zip files
-  fileslist <- list.files(paste0(getwd(), "/", destdir),
-                          recursive = T)
-  path_fileslist <- paste0(getwd(), "/", destdir, "/", fileslist)
-  path_pdffiles <- path_fileslist[grepl(".pdf", fileslist) == T]
-  unlink(path_pdffiles)
-  path_zipfiles <- path_fileslist[grepl(".zip", fileslist) == T]
-  unlink(path_zipfiles)
+  quarter <- match.arg(quarter)
+  type <- match.arg(type)
+  checkyear(year)
+  faers_link <- compose_faers_link(year, quarter, type)
+  faers_path <- compose_faers_path(path, year, quarter, type)
+  create_faers_folders(path, year, quarter, type)
+  filename <- paste0("faers_", type, "_", year, quarter, ".zip")
+  if(file.exists(paste0(faers_path, "/", filename))) {
+    stop(paste0("file ", filename, " already exists"))
+  } else {
+    if(utils::askYesNo(paste0("The file\n",
+                              filename,
+                              "\n will be downloaded at the path\n",
+                              faers_path,
+                              ".\nContinue?"))) {
+      cat("start downloading...\n")
+      downloader::download(faers_link,
+                           paste0(faers_path, "/", filename),
+                           mode = "wb")
+      cat("\ndone!")
+    }
+  }
 }